+++ native-client/assets/stylesheet.js
@@ -70,7 +70,8 @@ const styles = StyleSheet.create({
   inputFieldContainer: {
     flex: 1,
     justifyContent: 'center',
-    marginTop: 200
+    marginTop: 200,
+    alignItems: 'center'
   },
   inputFields: {
     width: 200
