+++ native-client/src/navigation/MainTabNavigator.js
@@ -4,7 +4,7 @@ import { Platform, Button } from 'react-native'
 import { Ionicons } from '@expo/vector-icons'
 import { StackNavigator, TabNavigator, TabBarBottom } from 'react-navigation'
 
-import { RecordScreen, ResultsScreen, Profile, SingleReport, AuthScreen, WordRepetition, Signup } from '../screens'
+import { RecordScreen, ResultsScreen, Profile, SingleReport, AuthScreen, WordRepetition, LoginScreen, SignupScreen } from '../screens'
 
 export default TabNavigator(
   {
