+++ native-client/src/screens/LoginScreen.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+import { View, Text, Button } from 'react-native'
+// import { ExpoLinksView } from '@expo/samples'
+import  {Login, Signup}  from '../components'
+import styles from '../../assets/stylesheet'
+
+function LoginScreen () {
+
+  return (
+    <Login />
+  )
+}
+
+export default LoginScreen
