+++ native-client/src/screens/SignupScreen.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+import { View, Text, Button } from 'react-native'
+// import { ExpoLinksView } from '@expo/samples'
+import  {Signup}  from '../components'
+import styles from '../../assets/stylesheet'
+
+function SignupScreen () {
+
+  return (
+    <Signup />
+  )
+}
+
+export default SignupScreen
