+++ native-client/src/screens/index.js
@@ -4,3 +4,5 @@ export { default as Profile } from './Profile.js'
 export { default as SingleReport } from './SingleReport.js'
 export { default as AuthScreen } from './AuthScreen.js'
 export { default as WordRepetition } from './WordRepetition.js'
+export {default as LoginScreen } from './LoginScreen'
+export {default as SignupScreen } from './SignupScreen'
